Add vtgate & vttablet support for "-mysql_auth_server_impl knox" "-db-credentials-server knox" respectively
Summary:
This should be used in combination with -knox_supported_usernames scriptro,longqueryro,...

The vtgate option will cause vtgate to authenticate users by checking the provided username/password against active credentials from knox.

The vttablet option will cause vttablet to act as the indicated user instead always of as a user named "vt_app" as we currently have it configured.

This should not be submitted until either the vtgate/vttablet code is updated to build with bazel to avoid breaking the jenkins build for vtgate. D255611 provides a useful bazel target. That should be landed, and then we'll want to update the jenkins config to user bazel instead of docker + make build.

Test Plan:
I manually tested the vtgate change.

I haven't actually tested the vttablet change yet. It's a bit more work to test a vttablet since it needs a mysql to point at, but I'll do that before landing probably by copying a binary to vittesttestingdb and starting a second tablet, then using grpc to sent a query.

Reviewers: laurabiester, emanuelecesena, esouhrada

Reviewed By: laurabiester

Subscribers: devinlundberg, #security

Differential Revision: https://phabricator.pinadmin.com/D255579