Support sending stats from vtgate or vttablet to OpenTSB
Summary:
Note: I doubt that this stat support handles timing histograms correctly, since it would try to put them all in a single stat and I'm guessing opentsdb requires them to be manually split into separate stats.

We can fix that in a subsequent diff. It would just be a matter of making sure that expvars of whatever types we want to support (like vitess timing/histogram) are translated into opentsdb Metric structs in the way we want.

Reviewers: laurabiester, vamsi, brian

Reviewed By: vamsi

Differential Revision: https://phabricator.pinadmin.com/D260218

Signed-off-by: David Weitzman <dweitzman@pinterest.com>