Parsing support for SHOW COLLATION WHERE
Summary:
This PR is on github at https://github.com/vitessio/vitess/pull/4067 and hopefully
helps with one of the issues noted in https://jira.pinadmin.com/browse/STERLING-4321

Both LIKE and WHERE are allowed in these queries:
https://dev.mysql.com/doc/refman/8.0/en/show-collation.html

Only adding WHERE for the moment because SQLAlchemy uses it in mysqldb.py when checking for UTF8 support.

Signed-off-by: David Weitzman <dweitzman@pinterest.com>

Test Plan: `./scripts/run_supported_tests.sh` passes on devapp. This is a subset of unit tests which includes the sqlparser and vtgate executor tests

Reviewers: laurabiester, vamsi

Reviewed By: laurabiester

Differential Revision: https://phabricator.pinadmin.com/D293562