Send grpc keepalive pings every 30s from vtgate on otherwise idle grpc connections.
Summary:
I don't know if this will accomplish anything, but it can't hurt. We observed some
kind of TCP broken connection error on vtgate and vttablet and I'm wondering if it
could be caused by the TCP connection getting killed at the OS level without vtgate
realizing it. Not sure if that fits all the symptoms, though.

Reviewers: anh, esouhrada

Reviewed By: esouhrada

Differential Revision: https://phabricator.pinadmin.com/D298085