Add a vindex that's like "hash", but allows an additive offset.
Summary:
This will be useful specifically for the "advertisers" table which only has
local IDs today during row inserts without any GID yet. Other tables will
have advertiser GIDs as they are created.

If we fully migrate to vitess sequences we can actually arrange to set both the
local and GID during the initial insert, and eventually we'll replace all the
local IDs with GIDs. It's possible we won't need this hash_offset. If not, we
can delete it later.

Reviewers: vamsi, sreshtav, chengchenghu, epruitt

Reviewed By: chengchenghu

Differential Revision: https://phabricator.pinadmin.com/D308766