add `VT_ARTIFACT` docker build abstraction for later use of `vtctld,vtworker`
Summary:
right now the single artifact `vitess` is effectively only `vtgate`, and we'd like to add `vtworker, vtctld` for the sake of `split shard` scripts/ui support.

as discussed with @dweitzman over slack. the build artifact essentially is a tar.gz with the right `telefig` in it. (the build must push `docker` images correctly for sure)
and the `builds.property` is the contract between `vitess` jenkins job and `publish` job, and we must include all artifacts we want

new `deploy` environment must be created to link to the `buildName` here to pick the right `telefig`, and that will pull the right docker image and start the process .

Test Plan:
http://jenkinsmaster-019.pinadmin.com:8080/job/vitess-more/configure

is created separated from the current `vitess` jenkins job. we'll use that to validate that `vtgate` is still published correctly, deployment is still ok.
after which point, we'll prepare the actual `telefig` for `vtworker & vtctld` and validate those artifacts are created and published as well.
then we'll add the corresponding deployment `env` to complete the loop (provisioning is needed i assume)

Reviewers: #vitess, dweitzman

Reviewed By: #vitess, dweitzman

Subscribers: jenkins

Differential Revision: https://phabricator.pinadmin.com/D341942