introducing `post_master_migration` hook, to be used when migration happens
Summary: for both vertical & horizontal migration, when the `master` shard flips, we're required to make some alignment, this needs to be reflected in daily dumps (determine which shards to dump from), incremental pipeline (which shards to subscribe to, with the help of maxwell). the hook introduced make the switch automated.

Reviewers: #vitess, epruitt, dweitzman

Reviewed By: #vitess, dweitzman

Subscribers: esouhrada, jenkins

Differential Revision: https://phabricator.pinadmin.com/D350742