Implement Create() for scatter_cache
Summary:
This should help in transactions where an entity is inserted
and then immediately updated, like a campaign. We update the
campaign spec ID and gid immediately after inserting it.

If we weren't able to target the write transaction using
the VitessTarget string in pepsi, at least we'll have a
good change of avoiding a scatter using scatter_cache

Reviewers: hzhou, #vitess, anh

Reviewed By: hzhou, #vitess

Subscribers: jenkins

Differential Revision: https://phabricator.pinadmin.com/D353129