Move knox role -> permission mapping and TLS cert checking logic into the common knox package
Summary:
Not at urgent diff, but this should make it easier to support the same authentication rules
(knox + cert checking) for the grpc vtgate interface.

That's useful because it means we could open to the grpc port on vtgates for convenience of
testing things and also potentially to use features that don't exist in the mysql API yet
like the update stream or event tokens.

Reviewers: #vitess, hzhou

Reviewed By: #vitess, hzhou

Subscribers: jenkins

Differential Revision: https://phabricator.pinadmin.com/D366005