Fix test with upstream merge
Summary:
The rebase on upstream changed the test result.
Check the difference here:
https://phabricator.pinadmin.com/D413269?vs=1525008&id=1527392&whitespace=ignore-most#toc

I need to fix the test.
Also file a ticket to run tests again before committing the code

Test Plan:
This was the failure
https://jenkinsmaster-019.pinadmin.com/job/vitess/411/console

Phabricator doesn't have a rule to rerun the test for now when code change happened, so it didn't catch that before

Reviewers: #vitess, hzhou

Reviewed By: #vitess, hzhou

Subscribers: jenkins

Differential Revision: https://phabricator.pinadmin.com/D414289