Mark sequence tables in vschema so "select next" queries work in v3 mode
Summary:
We want to be able to do "select next 10 values from goals_seq" in pepsi to pre-allocate IDs and GIDs,
since updating the GIDs after a bulk insert can be very expensive.

Technically we could use VitessTarget=patiogeneral:0 to do this with v2 mode, but it seems better
to properly annotate our vschema so it'll work in v3 mode.

Reviewers: #vitess, hzhou

Reviewed By: #vitess, hzhou

Subscribers: mingjianliu, jenkins

Differential Revision: https://phabricator.pinadmin.com/D422302