querylogs use `json` format instead
Summary:
the query logs are held in `vtgate` at the moment, so the change is mostly transparent.
but this would open up the chances to push this log to `elastic|sumologic` stack. and for us to have a better view of how the `vtgate` performs on a per query level

Reviewers: chengchenghu, O842 Vitess SOX, mingjianliu

Reviewed By: O842 Vitess SOX, mingjianliu

Subscribers: jenkins

Differential Revision: https://phabricator.pinadmin.com/D458880