enhancement of `validate-vschema`, now reports missing `sequence` table and `vindex` table
Summary:
[what]
this diff is an enhancement of the current `validate-vschema` used in `fix_vschema` tool for us to spot 2 more potential vschema problems:
1. missing `sequence `table from `DDLs`
2. missing `vindex` lookup table from `DDLs`

[why]
in the short term, this will help us spot schema issues in `mysql_change_management`, and lead to timely fixes of the gap if found.
longer term, we're in discussion with DBA team for the `skeema` based schema changes, and we'd like to offer a jenkins build which could take a full list of `DDLs` and apply `validate-vschema` for gate keeping the `DDLs` changes.

Reviewers: dweitzman, O842 Vitess SOX, mingjianliu

Reviewed By: dweitzman, O842 Vitess SOX, mingjianliu

Subscribers: epruitt, jenkins

Differential Revision: https://phabricator.pinadmin.com/D479078